requirements.txt
aws-shell>=0.2.2
boto3>=1.24.22
botocore>=1.27.28
joblib>=1.1.0
numpy>=1.21.6
ruamel.yaml>=0.17.21
ruamel.yaml.clib>=0.2.6
urllib3>=1.26.9
zstandard>0, <0.18.0
